Move dialect loading to main lib.
Dialects will not only be used by the thorin frontend, therefore they should be loadable from the thorin library.
