+++ cli/CMakeLists.txt
@@ -1,6 +1,4 @@
 add_executable(thorin
-    main.cpp
-    dialects.cpp
-    dialects.h)
+    main.cpp)
 
 target_link_libraries(thorin libthorin lyra)
